benchmark/systems/mediapipe/requirements.txt
mediapipe
Pillow
tqdm
numpy
opencv-python